+++ lib/ap/plan9/fcntl.c
@@ -12,6 +12,7 @@
 #include <errno.h>
 #include <stdarg.h>
 #include "sys9.h"
+#include <stdint.h>
 
 /*
  * BUG: advisory locking not implemented
@@ -25,7 +26,7 @@ fcntl(int fd, int cmd, ...)
 	int arg, i, ans, err;
 	Fdinfo *fi, *fans;
 	va_list va;
-	unsigned long oflags;
+	uint32_t oflags;
 
 	err = 0;
 	ans = 0;
